fix type specification, returning to the native types bool, int and float

+++ package/chemicalchecker/tool/targetmate/.old/old.py
@@ -287,8 +287,8 @@ class Old:
             neighs += [results[:k]]
         if N is None:
             N = len(query_smi)
-        sims = np.zeros((N, k), dtype=np.float32)
-        idxs = np.zeros((N, k), dtype=np.int)
+        sims = np.zeros((N, k), dtype=float )
+        idxs = np.zeros((N, k), dtype=int )
         for q_idx, hits in enumerate(neighs):
             sims_ = []
             idxs_ = []
